dynamic-plugins: rename to rpcplugin
