+++ internal/agent/agent_plugin.go
@@ -26,7 +26,7 @@ import (
 	"github.com/redpanda-data/benthos/v4/public/service"
 
 	agentruntimepb "github.com/redpanda-data/connect/v4/internal/agent/runtimepb"
-	"github.com/redpanda-data/connect/v4/internal/dynamic/plugin/runtimepb"
+	"github.com/redpanda-data/connect/v4/internal/rpcplugin/runtimepb"
 	"github.com/redpanda-data/connect/v4/internal/tracing"
 )
