+++ internal/cli/enterprise.go
@@ -20,9 +20,9 @@ import (
 
 	"github.com/redpanda-data/benthos/v4/public/service"
 
-	"github.com/redpanda-data/connect/v4/internal/dynamic/plugin"
 	"github.com/redpanda-data/connect/v4/internal/impl/kafka/enterprise"
 	"github.com/redpanda-data/connect/v4/internal/license"
+	"github.com/redpanda-data/connect/v4/internal/rpcplugin"
 	"github.com/redpanda-data/connect/v4/internal/telemetry"
 )
 
@@ -131,7 +131,7 @@ func InitEnterpriseCLI(binaryName, version, dateBuilt string, schema *service.Co
 				}
 
 				rpcPlugins := c.StringSlice("rpc-plugins")
-				err := plugin.DiscoverAndRegisterPlugins(service.OSFS(), service.GlobalEnvironment(), rpcPlugins)
+				err := rpcplugin.DiscoverAndRegisterPlugins(service.OSFS(), service.GlobalEnvironment(), rpcPlugins)
 				if err != nil {
 					return err
 				}
