+++ internal/rpcplugin/protogen.go
@@ -12,6 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package plugin
+package rpcplugin
 
-//go:generate protoc -I=../../../proto --go_opt=module=github.com/redpanda-data/connect/v4 --go-grpc_opt=module=github.com/redpanda-data/connect/v4 --go_out=../../.. --go-grpc_out=../../.. redpanda/runtime/v1alpha1/message.proto redpanda/runtime/v1alpha1/input.proto redpanda/runtime/v1alpha1/output.proto redpanda/runtime/v1alpha1/processor.proto
+//go:generate protoc -I=../../proto --go_opt=module=github.com/redpanda-data/connect/v4 --go-grpc_opt=module=github.com/redpanda-data/connect/v4 --go_out=../.. --go-grpc_out=../.. redpanda/runtime/v1alpha1/message.proto redpanda/runtime/v1alpha1/input.proto redpanda/runtime/v1alpha1/output.proto redpanda/runtime/v1alpha1/processor.proto
