+++ proto/redpanda/runtime/v1alpha1/output.proto
@@ -16,7 +16,7 @@ syntax = "proto3";
 
 package redpanda.runtime.v1alpha1;
 
-option go_package = "github.com/redpanda-data/connect/v4/internal/dynamic/plugin/runtimepb";
+option go_package = "github.com/redpanda-data/connect/v4/internal/rpcplugin/runtimepb";
 
 import "redpanda/runtime/v1alpha1/message.proto";
